refactor(em): decompose monolithic reconciler into 12 focused modules (#254)
Decomposes the 1868-line EffectivenessMonitor reconciler.go into 12 focused
files (all <500 lines) without changing observable behavior. The original
reconciler.go is now a thin 239-line orchestrator.

Extraction layout:
- scope.go: assessment scope determination, EA spec validation
- pod_health.go: active pod filtering, health stats computation
- target_resources.go: target resource queries, config hash resolution
- assess_components.go: health/hash/alert/metrics assessment logic
- completion.go: assessment completion, failure, metrics emission
- events.go: K8s event and audit event helpers
- reconcile_orchestrate.go: reconcileContext struct, main orchestration
- reconcile_validity_phase.go: WFP/stabilizing/expired/assessing phases
- reconcile_spec_drift.go: spec drift guard
- reconcile_components.go: component assessment pipeline
- reconcile_status.go: atomic status update and post-update events

REFACTOR improvements:
- Extract emitValidityExtendedEvent helper (eliminates 3x duplication)
- Safe hash truncation: min(23, len(hash)) prevents panic on short hashes
- Clarified comments in completion.go and reconcile_orchestrate.go

7 golden snapshot tests ensure behavioral equivalence across all phases.
All 347 EM unit tests pass with zero regressions.

BR-EM-254

Made-with: Cursor
